+++ HISTORY.rst
@@ -58,6 +58,7 @@
 - 1067_: [NetBSD] Process.cmdline() leaks memory if proces has terminated.
 - 1069_: [FreeBSD] Process.cpu_num() may return 255 for certain kernel
   processes.
+- 1074_: [FreeBSD] sensors_battery() raises OSError in case of no battery.
 
 **Porting notes**